modified and tested djed-sdk functions
