--- djed-sdk/src/coins/index.js
@@ -1,33 +1,0 @@
-export {
-    tradeDataPriceBuyRc,
-    tradeDataPriceSellRc,
-    buyRcTx,
-    sellRcTx,
-    checkBuyableRc,
-    checkSellableRc
-  } from "./reserveCoin";
-export {
-    tradeDataPriceBuySc,
-    tradeDataPriceSellSc,
-    buyScTx,
-    sellScTx,
-    checkBuyableSc,
-    checkSellableSc,
-    calculateFutureScPrice
-  } from "./stableCoin";
-  export {
-    scalingFactor,
-    FEE_UI_UNSCALED,
-    convertToBC,
-    tradeDataPriceCore,
-    calculateIsRatioBelowMax,
-    calculateIsRatioAboveMin,
-    isTxLimitReached,
-    promiseTx,
-    verifyTx,
-    calculateTxFees,
-    deductFees,
-    appendFees,
-    getFees
-  } from "./tradeUtils";
-
