+++ djed-sdk/src/index.js
@@ -1,3 +1,11 @@
-export { DjedInstance } from "./web3";
-export { getCoinDetails, getSystemParams, getAccountDetails, getCoinBudgets } from "./system";
-export * from "./coins/index";
+export { getWeb3 } from "./web3";
+export {
+  calculateBcUsdEquivalent,
+  calculateRcUsdEquivalent,
+  getBcUsdEquivalent,
+  getRcUsdEquivalent,
+  getScAdaEquivalent,
+} from "./helpers";
+
+export * from "./djed/index";
+export * from "./oracle";
