+++ djed-sdk/src/oracle/index.js
@@ -1,0 +1,1 @@
+export { getOracleAddress, getOracleContract } from "./oracle";
